reverting datatype change to be able to release this change before version 3.0

+++ src/main/java/redis/clients/jedis/BasicCommands.java
@@ -34,7 +34,7 @@ public interface BasicCommands {
 
   String slaveofNoOne();
 
-  int getDB();
+  Long getDB();
 
   String debug(DebugParams params);
