+++ src/main/java/redis/clients/jedis/BinaryJedis.java
@@ -2857,7 +2857,7 @@ public class BinaryJedis implements BasicCommands, BinaryJedisCommands, MultiKey
     }
   }
 
-  public int getDB() {
+  public Long getDB() {
     return client.getDB();
   }
